+++ src/data/runtime.lex/2016-11-28/examples-1.json.php
@@ -1,0 +1,3 @@
+<?php
+// This file was auto-generated from sdk-root/src/data/runtime.lex/2016-11-28/examples-1.json
+return [ 'version' => '1.0', 'examples' => [],];